Implement authorization using Laravel Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -69,6 +69,14 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        //if (\Gate::allows('update-question', $question)) {
+            // move to view
+        //}
+        // or
+        if (\Gate::denies('update-question', $question)) {
+            abort('403', 'Access denied.');
+        }
+
         return view('questions.edit', compact('question'));
     }
 
@@ -81,6 +89,14 @@ class QuestionsController extends Controller
      */
     public function update(Requests\AskQuestionRequest $request, Question $question)
     {
+        //if (\Gate::allows('update-question', $question)) {
+            // move to view
+        //}
+        // or
+        if (\Gate::denies('update-question', $question)) {
+            abort('403', 'Access denied.');
+        }
+
         $question->update($request->except('_token'));
 
         return redirect()->route('questions.index')->with('success', 'Your question has been updated.');
@@ -94,6 +110,14 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        //if (\Gate::allows('delete-question', $question)) {
+            // execute delete
+        //}
+        // or
+        if (\Gate::denies('delete-question', $question)) {
+            abort('403', 'Access denied.');
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Question has been deleted.');
